add dtype to normalize function

+++ ponyo/pipeline.py
@@ -146,7 +146,7 @@ def normalize_expression_data(
     """
 
     # Read data
-    data = pd.read_csv(raw_input_data_file, header=0, sep="\t", index_col=0)
+    data = pd.read_csv(raw_input_data_file, header=0, sep="\t", index_col=0, dtype=float)
 
     print(
         "input: dataset contains {} samples and {} genes".format(
